boardconfig: Remove generic variant target
This change was apparently inspired by a commit to the Pixel 3
LOS device tree, which states that they want to make the system
image more generic.

Well, I'm not sure what that is about, but in this case I'm
wondering what other devices are going to use the roms we build?

Not going to happen ever, so let's not compromise on compiler
optimization for this impossible scenario.